+++ LT_Algo_Module/my-app/src/main.js
@@ -2,6 +2,12 @@ const { app, BrowserWindow } = require('electron');
 const { maxHeaderSize } = require('http');
 const path = require('path');
 
+//import custom classes
+const Vector2 = require('./vector2.js')
+const Vector3 = require('./vector3.js')
+const Matrix2x2 = require('./matrix2x2.js')
+const Plane = require('./plane.js')
+
 // Handle creating/removing shortcuts on Windows when installing/uninstalling.
 if (require('electron-squirrel-startup')) {
   // eslint-disable-line global-require
@@ -56,115 +62,13 @@ app.on('activate', () => {
 // In this file you can include the rest of your app's specific main process
 // code. You can also put them in separate files and import them here.
 
-class Vector2 {
-  constructor(x,y){
-    this.x = x
-    this.y = y
-  }
-  matMult(a,b,c,d) {
-    let newx = a*this.x+b*this.y
-    let newy = c*this.x+d*this.y
 
-    this.x = newx
-    this.y = newy
-  }
-}
-class Vector3 {
-  constructor(x,y,z){
-    this.x = x
-    this.y = y
-    this.z = z
-  }
-}
-/*2x2 matrix class in the form of:
- [a b; c d]
-*/
-class Matrix2x2 {
-  constructor(a,b,c,d){
-    this.a = a
-    this.b = b
-    this.c = c
-    this.d = d
-  }
-  invert(){
-    let det = (1/(this.a*this.d-this.b*this.c))
-    det = det.toFixed(3)
-    let tempA = this.d*det
-    let tempD = this.a*det
-    this.a = tempA
-    this.b = -1*this.b*det
-    this.c = -1*this.c*det
-    this.d = tempD
-    this.a = this.a.toFixed(3)
-    this.b = this.b.toFixed(3)
-    this.c = this.c.toFixed(3)
-    this.d = this.d.toFixed(3)
-  }
-}
 
-//plane class, takes in 3 vectors and calculates the plane
-//v1 is origin, v2 is i hat, v3 is j hat
-//distance data comes in, corner sensor goes to v1 <0,0,data1>
-//other bottom sensor goes to v2 <1,0,data2>
-//top sensor goes to v3 <0,1,data3>
-//v1-v3 are points on the projection plane treated as vectors from the origin ('center' of bulb), for the sake of calculations
-class Plane {
-  constructor(v1,v2,v3) {
-    this.v1 = v1
-    this.v2 = v2
-    this.v3 = v3
-    var equation = [0,0,0,0]//ax+by+cz=d
-  }
-  calcEq(){
-    //vectors in the plane
-    let vp1 = new Vector3(this.v1.x-this.v2.x,this.v1.y-this.v2.y,this.v1.z-this.v2.z)
-    let vp2 = new Vector3(this.v1.x-this.v3.x,this.v1.y-this.v3.y,this.v1.z-this.v3.z)
-    //cross product of vectors in the plane
-    let xProd = new Vector3(vp1.y*vp2.z-vp1.z*vp2.y,-1*(vp1.x*vp2.z-vp1.z*vp2.x),vp1.x*vp2.y-vp1.y*vp2.x)
-    //create plane equation, RHS being the constant on the right hand side
-    let RHS = this.v1.x*xProd.x + this.v1.y*xProd.y + this.v1.z*xProd.z
-    //equation is the cross product <x,y,z> = RHS
-    this.equation = [xProd.x,xProd.y,xProd.z,RHS]
-  }
 
-  //unit vectors determined from sensors (v2 and v3) are hard coded to <1,0,dist> and <0,1,dist>
-  //these methods use the orign (v1) and the sensors do depermine the actual distance between the origin and the unit sensors
-  //iHat and jHat are used to determine transformation
-  getIHat(){
-    let numx = Math.pow(this.v2.x - this.v1.x,2)
-    let numy = Math.pow(this.v2.y - this.v1.y,2)
-    let numz = Math.pow(this.v2.z - this.v1.z,2)
-    let dist = Math.sqrt(numx+numy+numz)
-    let theta = this.getAngle()
-    let horz = dist*Math.cos(theta)
-    let vert = dist*Math.sin(theta)
-    let iHat = new Vector2(horz,vert)
-    return iHat
-  }
 
-  getJHat(){
-    let numx = Math.pow(this.v3.x - this.v1.x,2)
-    let numy = Math.pow(this.v3.y - this.v1.y,2)
-    let numz = Math.pow(this.v3.z - this.v1.z,2)
-    let dist = Math.sqrt(numx+numy+numz)
-    let theta = this.getAngle()
-    let horz = dist*Math.cos(theta)
-    let vert = dist*Math.sin(theta)
-    let jHat = new Vector2(vert,horz)
-    return jHat
-  }
-  
-  //calculates the angle between the projection plane and a plane that would be normal to the projector 0x+0y+1z=0
-  getAngle() {
-    //numerator = |A1*A2+B1*B2+C1*C2| = |A1*0+B1*0+C1*1|
-    let numerator = Math.abs(this.equation[2])
-    //denominator = sqrt(A1^2+A2^2+A3^2)*sqrt(0^2+0^2+1^2)
-    let denominator = Math.sqrt(Math.pow(this.equation[0],2)+Math.pow(this.equation[1],2)+Math.pow(this.equation[2],2))
-    let cosTheta = numerator/denominator
-    let theta = Math.acos(cosTheta)
-    return theta
-  }
-}
+
+
+
 
 var vec1 = new Vector3(0,0,5)
 var vec2 = new Vector3(1,0,6)
@@ -207,8 +111,6 @@ console.log('Vector2: ')
 console.log(newVec2)
 console.log('Vector3: ')
 console.log(newVec3)
-
-
