+++ LT_Algo_Module/my-app/src/vector3.js
@@ -1,0 +1,8 @@
+class Vector3 {
+    constructor(x,y,z){
+      this.x = x
+      this.y = y
+      this.z = z
+    }
+  }
+  module.exports = Vector3
